// src/types/talk.ts
export interface SpeakerFormData {
    fullName: string;
    email: string;
    phone: string;
    company: string;
    title: string;
    bio: string;
    talkTitle: string;
    talkDescription: string;
    talkDuration: string;
    talkLevel: string;
    previousSpeakingExperience: string;
    additionalNotes: string;
  }